+++ proxy-server/utils/StandingsUtils.js
@@ -1,4 +1,4 @@
-import { HIGH_TO_LOW } from "../NewConstants.js";
+import { HIGH_TO_LOW } from "../APIConstants.js";
 import pick from "lodash/pick.js";
 import sum from "lodash/sum.js";
 import values from "lodash/values.js";
